move path utils
